requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.25
langchain-community==0.3.25
langgraph==0.4.1
langgraph-checkpoint-sqlite==2.0.6
qdrant-client==1.12.1
sentence-transformers==3.3.1
fastmcp==2.5.1
prometheus-client==0.21.0
httpx==0.28.1
apscheduler==3.10.4
pydantic-settings==2.6.1
